Address FIX-82 QA findings
