+++ crates/atm/tests/send.rs
@@ -454,7 +454,7 @@ fn test_send_runs_post_send_hook_with_expected_payload() {
     assert!(payload["message_id"].as_str().is_some());
     assert!(payload.get("task_id").is_some());
     assert_eq!(payload["hook_match"]["sender"], true);
-    assert_eq!(payload["hook_match"]["recipient"], false);
+    assert_eq!(payload["hook_match"]["recipient"], true);
 }
 
 #[test]
@@ -506,9 +506,36 @@ fn test_send_emits_post_send_hook_skip_warning_when_no_filter_matches() {
         fixture.stderr(&output)
     );
     assert!(!payload_path.exists(), "hook payload unexpectedly created");
-    assert!(fixture.stderr(&output).contains("post-send hook skipped"));
+    assert_eq!(
+        fixture.stderr(&output),
+        "post-send hook skipped: sender arch-ctm not in post_send_hook_senders team-lead\nand recipient recipient not in post_send_hook_recipients quality-mgr\n"
+    );
     let inbox = fixture.inbox_contents("recipient");
     assert_eq!(inbox.len(), 1);
+}
+
+#[test]
+fn test_send_emits_post_send_hook_skip_warning_on_stderr_in_json_mode() {
+    let fixture = Fixture::new("recipient");
+    let (hook_path, payload_path) = fixture.install_hook_fixture("capture");
+    fixture.write_atm_config(&format!(
+        "[atm]\npost_send_hook = ['{}', 'capture', '{}']\npost_send_hook_senders = ['team-lead']\npost_send_hook_recipients = ['quality-mgr']\n",
+        hook_path.display(),
+        payload_path.display()
+    ));
+
+    let output = fixture.run(&["send", "recipient@atm-dev", "hello skipped hook", "--json"]);
+
+    assert!(
+        output.status.success(),
+        "stderr: {}",
+        fixture.stderr(&output)
+    );
+    assert!(!payload_path.exists(), "hook payload unexpectedly created");
+    assert_eq!(
+        fixture.stderr(&output),
+        "post-send hook skipped: sender arch-ctm not in post_send_hook_senders team-lead\nand recipient recipient not in post_send_hook_recipients quality-mgr\n"
+    );
 }
 
 #[test]
@@ -530,7 +557,7 @@ fn test_send_runs_post_send_hook_when_recipient_matches_filter() {
     );
     let payload: serde_json::Value =
         serde_json::from_slice(&fs::read(payload_path).expect("hook payload")).expect("json");
-    assert_eq!(payload["hook_match"]["sender"], false);
+    assert_eq!(payload["hook_match"]["sender"], true);
     assert_eq!(payload["hook_match"]["recipient"], true);
 }
 
@@ -629,7 +656,7 @@ fn test_send_post_send_hook_receives_only_configured_positional_args() {
     assert_eq!(captured["args"], serde_json::json!([]));
     assert_eq!(captured["payload"]["to"], "recipient@atm-dev");
     assert_eq!(captured["payload"]["hook_match"]["sender"], true);
-    assert_eq!(captured["payload"]["hook_match"]["recipient"], false);
+    assert_eq!(captured["payload"]["hook_match"]["recipient"], true);
 }
 
 #[test]
@@ -652,7 +679,7 @@ fn test_send_runs_post_send_hook_when_sender_filter_is_wildcard() {
     let payload: serde_json::Value =
         serde_json::from_slice(&fs::read(payload_path).expect("hook payload")).expect("json");
     assert_eq!(payload["hook_match"]["sender"], true);
-    assert_eq!(payload["hook_match"]["recipient"], false);
+    assert_eq!(payload["hook_match"]["recipient"], true);
 }
 
 #[test]
@@ -674,7 +701,30 @@ fn test_send_runs_post_send_hook_when_recipient_filter_is_wildcard() {
     );
     let payload: serde_json::Value =
         serde_json::from_slice(&fs::read(payload_path).expect("hook payload")).expect("json");
-    assert_eq!(payload["hook_match"]["sender"], false);
+    assert_eq!(payload["hook_match"]["sender"], true);
+    assert_eq!(payload["hook_match"]["recipient"], true);
+}
+
+#[test]
+fn test_send_runs_post_send_hook_when_filter_lists_are_empty() {
+    let fixture = Fixture::new("recipient");
+    let (hook_path, payload_path) = fixture.install_hook_fixture("capture");
+    fixture.write_atm_config(&format!(
+        "[atm]\npost_send_hook = ['{}', 'capture', '{}']\n",
+        hook_path.display(),
+        payload_path.display()
+    ));
+
+    let output = fixture.run(&["send", "recipient@atm-dev", "hello empty filters"]);
+
+    assert!(
+        output.status.success(),
+        "stderr: {}",
+        fixture.stderr(&output)
+    );
+    let payload: serde_json::Value =
+        serde_json::from_slice(&fs::read(payload_path).expect("hook payload")).expect("json");
+    assert_eq!(payload["hook_match"]["sender"], true);
     assert_eq!(payload["hook_match"]["recipient"], true);
 }
 
@@ -690,9 +740,8 @@ fn test_send_rejects_retired_post_send_hook_members_config() {
     assert!(!output.status.success());
     let stderr = fixture.stderr(&output);
     assert!(stderr.contains("post_send_hook_members"));
-    assert!(stderr.contains("post_send_hook_senders"));
-    assert!(stderr.contains("post_send_hook_recipients"));
-    assert!(stderr.contains("Use '*' to match all senders or all recipients."));
+    assert!(stderr.contains(".atm.toml"));
+    assert!(stderr.contains("Replace post_send_hook_members with post_send_hook_senders and/or post_send_hook_recipients under [atm]. Use * to match all."));
 }
 
 #[test]
